refactor: move tests to internal package and add long test support
