+++ vm_test.go
@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/ozanh/ugo/tests"
+	"github.com/ozanh/ugo/internal/tests"
 
 	. "github.com/ozanh/ugo"
 )
